Splitting up conftest to make both pytest happy and pick up the configuration properly with tox
